+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## Change log
 ----------------------
 
-Version 2-SNAPSHOT
+Version 2
 -------------
 
 ADDED: